requirements.txt
aiogram>=3.14.0
python-dotenv>=1.0.1
google-api-python-client>=2.146.0
google-auth>=2.31.0
